backend env fix, image-transition, renaming & cleanup
