+++ immichFrame.Web/src/lib/components/elements/image-component.svelte
